Start on turn costs. Implement edge angles.

+++ src/algos/edge_router.h
@@ -304,7 +304,7 @@ class EdgeRouter {
         WaySharedAttrs::Create({.dir = 1, .access = ACC_YES, .maxspeed = 120});
     return ctx.metric.Compute(
         g_wsa, ctx.opt.vt, DIR_FORWARD,
-        {.distance_cm = static_cast<uint64_t>(
+        {.distance_cm = static_cast<uint32_t>(
              1.00 * calculate_distance(node.lat, node.lon, ctx.target_lat,
                                        ctx.target_lon)),
          .contra_way = 0});
